requirements.txt
langchain_core==1.2.20
langchain_openai==1.1.11
langgraph==1.1.3
markdown-it-py==3.0.0
python-dotenv==1.2.2
Requests==2.32.5
streamlit==1.50.0
weasyprint==68.1
